moved function out of tools.js
